# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateOrderRequest:
    properties:
      listing_id:
        type: string
      qty:
        type: integer
    type: object
  handlers.CreateRestaurantRequest:
    properties:
      address:
        type: string
      closing_time:
        description: 'Format: "HH:MM:SS"'
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
    type: object
  handlers.UpdateOrderStatusRequest:
    properties:
      status:
        description: '"pending", "ready", "completed", "cancelled"'
        type: string
    type: object
...
  models.Listing:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        description: Nullable for mystery box
        type: string
      orders:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      photo_url:
        type: string
      pickup_time:
        $ref: '#/definitions/models.TimeOnly'
      price:
        description: Price in smallest currency unit (e.g., cents)
        type: integer
      restaurant:
        allOf:
        - $ref: '#/definitions/models.Restaurant'
        description: Relationships
      restaurant_id:
        type: string
      stock:
        type: integer
      type:
        $ref: '#/definitions/models.ListingType'
    type: object
  models.ListingType:
    enum:
    - mystery_box
    - reveal
    type: string
    x-enum-varnames:
    - ListingTypeMysteryBox
    - ListingTypeReveal
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: string
      listing:
        $ref: '#/definitions/models.Listing'
      listing_id:
        type: string
      qty:
        type: integer
      status:
        $ref: '#/definitions/models.OrderStatus'
      total_price:
        description: Total price in smallest currency unit
        type: integer
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.OrderStatus:
    enum:
    - pending
    - ready
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusReady
    - OrderStatusCompleted
    - OrderStatusCancelled
  models.Restaurant:
    properties:
      address:
        type: string
      closing_time:
        $ref: '#/definitions/models.TimeOnly'
      created_at:
        type: string
      id:
        type: string
      lat:
        type: number
      listings:
        items:
          $ref: '#/definitions/models.Listing'
        type: array
      lng:
        type: number
      name:
        type: string
      owner:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      owner_id:
        type: string
    type: object
  models.TimeOnly:
    properties:
      time.Time:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      orders:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      restaurants:
        description: Relationships
        items:
          $ref: '#/definitions/models.Restaurant'
        type: array
      role:
        $ref: '#/definitions/models.UserRole'
    type: object
  models.UserRole:
    enum:
    - user
    - restaurant
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleRestaurant
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@eatright.com
    name: API Support
  description: Production-ready backend for EatRight food waste reduction platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EatRight API
  version: "1.0"
paths:
...
  /orders:
    post:
      consumes:
      - application/json
      description: Creates a new food order and automatically decrements stock
      parameters:
      - description: Order Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Order created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Order'
              type: object
        "400":
          description: Invalid request or insufficient stock
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - Orders
  /orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: Updates order status - pending/ready/completed/cancelled (restaurant
        owner only)
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Status Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Order status updated successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden - not restaurant owner
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - Orders
  /orders/me:
    get:
      consumes:
      - application/json
      description: Retrieves all orders placed by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Orders retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Order'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get user's order history
      tags:
      - Orders
  /restaurants:
    get:
      consumes:
      - application/json
      description: Retrieves all restaurants or nearby restaurants if lat/lng provided
      parameters:
      - description: Latitude for nearby search
        in: query
        name: lat
        type: number
      - description: Longitude for nearby search
        in: query
        name: lng
        type: number
      - description: 'Max distance in km (default: 10)'
        in: query
        name: distance
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Restaurants retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Restaurant'
                  type: array
              type: object
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List restaurants
      tags:
      - Restaurants
    post:
      consumes:
      - application/json
      description: Creates a new restaurant (requires restaurant role)
      parameters:
      - description: Restaurant Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRestaurantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Restaurant created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Restaurant'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden - restaurant role required
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create restaurant
      tags:
      - Restaurants
  /restaurants/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific restaurant
      parameters:
      - description: Restaurant ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Restaurant retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Restaurant'
              type: object
        "400":
          description: Invalid restaurant ID
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Restaurant not found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get restaurant by ID
      tags:
      - Restaurants
...
  /users/me:
    get:
      consumes:
      - application/json
      description: Retrieves the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"